+++ CHANGES.md
@@ -12,16 +12,41 @@ of that change.
 
 
 
-
-## 1.0.17.dev0 (Work In Progress)
+## 1.0.19.dev0 (Work In Progress)
 
 ### New:
 
+- add an argument `ResizeMtd` that tells `apply_tfms` how to resize the image to the desired size (crop, pad, squish or no).
+- all the image dataset have an `image_opener` attribute (default `open_image`) that can be changed. The `SegmentationDataset` has a `mask_opener` attribute.
+
 ### Changed:
+
+- `verify_images` can know resize images on top of checking they're not broken.
 
 ### Fixed:
 
+- `ImageDataBunch.from_df` doesn't change the dataframe.
 
+## 1.0.18 (2018-10-30)
+
+### Fixed:
+
+- Fix jupyter dep version
+
+
+## 1.0.17 (2018-10-30)
+
+### New:
+
+- Add tiny datasets
+
+### Changed:
+
+- remove wrong `Fbeta`
+
+### Fixed:
+
+- fix implementation of `fbeta`
 
 ## 1.0.16 (2018-10-30)
 
@@ -35,7 +60,6 @@ of that change.
 - Monkey-patched torch.Tensor so matplotlib works
 - `Learner.create_unet`
 - Data block API
-- Add tiny datasets
 
 ### Changed: